+++ dh_tools/Cargo.toml
@@ -9,6 +9,3 @@ edition = "2021"
 aes_tools = { path = "../aes_tools" }
 x25519-dalek = "2.0.0-pre.1"
 rand_core = "0.6.3"
-
-[dev-dependencies]
-fennel-lib = { path = "../" }
